{{cookiecutter.project_slug}}/requirements_dev.txt
-r requirements.txt

scrapyd-client==1.1.0
